pyproject.toml
[project]
name = "nano-banana-studio"
version = "1.1.0"
description = "A studio for generating images via Gemini API with full parameter control."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "PyQt6>=6.10.1",
    "Pillow>=12.0.0",
    "requests>=2.32.5",
    "pydantic>=2.12.5",
    "google-genai>=1.56.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
